X11: Always save screen number as usize
The screen number is only ever used for indexing into the list of
screens. Thus, usize is the perfect type for this. x11rb even provides
the screen number as usize.

This commit thus can get rid of a couple of casts by just always using
usize.

(History guess: xcb-rs uses i32 for the screen number (because C) and
that's why the code was like this in the first place)

+++ druid-shell/src/platform/x11/screen.rs
@@ -37,7 +37,7 @@ where
 pub(crate) fn get_monitors() -> Vec<Monitor> {
     let result = if let Some(app) = crate::Application::try_global() {
         let app = app.platform_app;
-        get_monitors_impl(app.connection().as_ref(), app.screen_num() as usize)
+        get_monitors_impl(app.connection().as_ref(), app.screen_num())
     } else {
         let (conn, screen_num) = match x11rb::connect(None) {
             Ok(res) => res,
